move QmlSettings to separate file
