+++ libosmscout-client-qt/include/osmscoutclientqt/QtSettingsStorage.h
@@ -1,5 +1,5 @@
-#ifndef LIBOSMSCOUT_QTSETTINGSSTORAGE_H
-#define LIBOSMSCOUT_QTSETTINGSSTORAGE_H
+#ifndef OSMSCOUT_CLIENT_QT_QTSETTINGSSTORAGE_H
+#define OSMSCOUT_CLIENT_QT_QTSETTINGSSTORAGE_H
 
 /*
   OSMScout - a Qt backend for libosmscout and libosmscout-map
@@ -57,4 +57,4 @@ namespace osmscout {
   };
 
 }
-#endif //LIBOSMSCOUT_QTSETTINGSSTORAGE_H
+#endif //OSMSCOUT_CLIENT_QT_QTSETTINGSSTORAGE_H
